requirements.txt
streamlit==1.30.0
pandas==1.5.3
streamlit-option-menu==0.3.6
scikit-learn==1.3.0
numpy==1.26.4
tensorflow==2.17.0
torch==2.1.1
torchvision==0.16.1
plotly==5.18.0
matplotlib==3.10.0
seaborn==0.12.2
pillow